add figures tab
